feat: 🎸 connect api to add-room
