+++ src/pages/index.tsx
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { createStore } from 'store';
 
-import AddRoom from './add-room';
-import Manual from './add-room/manual';
+import AddRoom from './add-room/AddRoom';
+import AddRoomViaLink from './add-room/via-link';
 import Login from './auth/Login';
 import ChecklistPage from './checklist/ChecklistPage';
 import RoomComparisonPage from './comparison/RoomComparison';
@@ -29,7 +29,7 @@ export default function EntryRoute() {
           <PrivateRoute path='/comparison' component={RoomComparisonPage} />
           <PrivateRoute path='/auth' component={Login} />
           <PrivateRoute path='/rooms' component={RoomList} />
-          <Route path='/add-room/manual' component={Manual} />
+          <Route path='/add-room/via-link' component={AddRoomViaLink} />
           <Route path='/add-room' component={AddRoom} />
           <Route path='/' component={MainPage} />
         </Switch>
